# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hospital CRM & Patient Engagement Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contact-variables:
    get:
      operationId: listContactVariables
      tags: [contact-variables]
      summary: List contact variables
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactVariable"
    post:
      operationId: createContactVariable
      tags: [contact-variables]
      summary: Create a contact variable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactVariableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactVariable"
  /contact-variables/{id}:
    patch:
      operationId: updateContactVariable
      tags: [contact-variables]
      summary: Update a contact variable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactVariableUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactVariable"
    delete:
      operationId: deleteContactVariable
      tags: [contact-variables]
      summary: Delete a contact variable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Leads ─────────────────────────────────────────────────────────────────
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/submit:
    post:
      operationId: submitCampaign
      tags: [campaigns]
      summary: Submit campaign for manager approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/approve:
    post:
      operationId: approveCampaign
      tags: [campaigns]
      summary: Approve campaign — runs compliance gate, debits wallet, goes live
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approval result with compliance checks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignApprovalResult"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a live campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paused campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
components:
  schemas:
...
    ContactVariable:
      type: object
      required: [id, name, description, mandatory, system]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        mandatory:
          type: boolean
        system:
          type: boolean
    ContactVariableInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        mandatory:
          type: boolean
    ContactVariableUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        mandatory:
          type: boolean

    # ── Lead ─────────────────────────────────────────────────────────────────
...
    CampaignChannel:
      type: object
      required: [channel, templateId, perMessageCost]
      properties:
        channel:
          type: string
          enum: [sms, whatsapp, push]
        templateId:
          type: integer
        templateName:
          type: ["string", "null"]
        perMessageCost:
          type: number
        senderId:
          type: ["string", "null"]
        wabaContact:
          type: ["string", "null"]
...
    ComplianceCheck:
      type: object
      required: [name, passed, message]
      properties:
        name:
          type: string
        passed:
          type: boolean
        message:
          type: string
        droppedCount:
          type: ["integer", "null"]
    Campaign:
      type: object
      required: [id, name, goal, audienceSegmentId, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        goal:
          type: string
        audienceSegmentId:
          type: integer
        audienceSegmentName:
          type: ["string", "null"]
        channels:
          type: array
          items:
            $ref: "#/components/schemas/CampaignChannel"
        estimatedRecipients:
          type: ["integer", "null"]
        costBreakdown:
          type: ["object", "null"]
          properties:
            channelCost:
              type: number
            fee:
              type: number
            gst:
              type: number
            total:
              type: number
        status:
          type: string
          enum: [draft, submitted, approved, live, paused, completed, failed]
        complianceResult:
          type: ["object", "null"]
          properties:
            checks:
              type: array
              items:
                $ref: "#/components/schemas/ComplianceCheck"
            passed:
              type: boolean
        createdBy:
          type: integer
        createdByName:
          type: ["string", "null"]
        approvedBy:
          type: ["integer", "null"]
        approvedByName:
          type: ["string", "null"]
        createdAt:
          type: string
        launchedAt:
          type: ["string", "null"]
    CampaignInput:
      type: object
      required: [name, goal, audienceSegmentId]
      properties:
        name:
          type: string
        goal:
          type: string
        audienceSegmentId:
          type: integer
        channels:
          type: array
          items:
            $ref: "#/components/schemas/CampaignChannel"
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
        audienceSegmentId:
          type: integer
        channels:
          type: array
          items:
            $ref: "#/components/schemas/CampaignChannel"
    CampaignApprovalResult:
      type: object
      required: [campaign, complianceChecks, passed]
      properties:
        campaign:
          $ref: "#/components/schemas/Campaign"
        complianceChecks:
          type: array
          items:
            $ref: "#/components/schemas/ComplianceCheck"
        passed:
          type: boolean
        invoice:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            number:
              type: string
            amount:
              type: number
            gst:
              type: number
...